+++ packages/cloud_functions/cloud_functions_platform_interface/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 5.1.29
+
+ - Update a dependency to the latest release.
+
+## 5.1.28
+
+ - Update a dependency to the latest release.
+
+## 5.1.27
+
+ - Update a dependency to the latest release.
+
 ## 5.1.26
 
  - Update a dependency to the latest release.