+++ packages/firebase_auth/firebase_auth_web/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 5.2.7
+
+ - Update a dependency to the latest release.
+
+## 5.2.6
+
+ - Update a dependency to the latest release.
+
+## 5.2.5
+
+ - Update a dependency to the latest release.
+
 ## 5.2.4
 
  - Update a dependency to the latest release.