+++ packages/firebase_dynamic_links/firebase_dynamic_links_platform_interface/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.2.3+29
+
+ - Update a dependency to the latest release.
+
+## 0.2.3+28
+
+ - Update a dependency to the latest release.
+
+## 0.2.3+27
+
+ - Update a dependency to the latest release.
+
 ## 0.2.3+26
 
  - Update a dependency to the latest release.