+++ packages/firebase_messaging/firebase_messaging_web/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 3.2.14
+
+ - Update a dependency to the latest release.
+
+## 3.2.13
+
+ - Update a dependency to the latest release.
+
+## 3.2.12
+
+ - Update a dependency to the latest release.
+
 ## 3.2.11
 
  - Update a dependency to the latest release.