+++ packages/firebase_ui_oauth_apple/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 1.0.20
+
+ - Update a dependency to the latest release.
+
+## 1.0.19
+
+ - Update a dependency to the latest release.
+
+## 1.0.18
+
+ - Update a dependency to the latest release.
+
+## 1.0.17
+
+ - Update a dependency to the latest release.
+
 ## 1.0.16
 
  - Update a dependency to the latest release.